+++ src/lib.rs
@@ -109,6 +109,7 @@ pub use crate::algorithms::constructors::famous::{famous, famous_names};
 pub use crate::algorithms::constructors::full::full_graph;
 pub use crate::algorithms::constructors::generalized_petersen::generalized_petersen;
 pub use crate::algorithms::constructors::hamming::hamming;
+pub use crate::algorithms::constructors::hexagonal_lattice::hexagonal_lattice;
 pub use crate::algorithms::constructors::hypercube::{MAX_HYPERCUBE_DIMENSION, hypercube};
 pub use crate::algorithms::constructors::kary_tree::{TreeMode, kary_tree};
 pub use crate::algorithms::constructors::kautz::kautz;